+++ tests/unit/test_tasks/test_psfgen_preserve/00-01-000_psfgen-build.tcl
@@ -1,0 +1,192 @@
+# pestifer.scripters: 00-01-000_psfgen-build.tcl
+####################### Created Wed Aug 19 09:53:10 2026 #######################
+package require psfgen
+psfcontext mixedcase
+topology top_all36_prot.rtf
+topology top_all36_cgenff.rtf
+topology top_all36_lipid.rtf
+topology top_all36_carb.rtf
+topology top_all36_na.rtf
+topology toppar_water_ions.str
+topology toppar_all36_carb_glycopeptide.str
+topology toppar_all36_carb_imlab.str
+topology toppar_all36_prot_modify_res.str
+topology toppar_all36_moreions.str
+topology top_all35_ethers.rtf
+pdbalias atom ILE CD1 CD
+pdbalias atom MET SE SD
+pdbalias atom BGLCNA C7 C
+pdbalias atom BGLCNA O7 O
+pdbalias atom BGLCNA C8 CT
+pdbalias atom BGLCNA N2 N
+pdbalias atom ANE5AC C10 C
+pdbalias atom ANE5AC C11 CT
+pdbalias atom ANE5AC N5 N
+pdbalias atom ANE5AC O1A O11
+pdbalias atom ANE5AC O1B O12
+pdbalias atom ANE5AC O10 O
+pdbalias atom VCG C01 C1
+pdbalias atom VCG C02 C2
+pdbalias atom VCG C03 C3
+pdbalias atom VCG C04 C4
+pdbalias atom VCG C05 C5
+pdbalias atom VCG C06 C6
+pdbalias atom VCG C07 C7
+pdbalias atom VCG C08 C8
+pdbalias atom VCG C09 C9
+pdbalias atom TIP3 O OH2
+pdbalias atom CLA CL CLA
+pdbalias atom ACT CH3 C1
+pdbalias atom ACT O O1
+pdbalias atom ACT OXT O2
+pdbalias atom ACT C C2
+pdbalias atom CAL CA CAL
+pdbalias atom ILE CD1 CD
+pdbalias atom MET SE SD
+pdbalias atom BGLCNA C7 C
+pdbalias atom BGLCNA O7 O
+pdbalias atom BGLCNA C8 CT
+pdbalias atom BGLCNA N2 N
+pdbalias atom ANE5AC C10 C
+pdbalias atom ANE5AC C11 CT
+pdbalias atom ANE5AC N5 N
+pdbalias atom ANE5AC O1A O11
+pdbalias atom ANE5AC O1B O12
+pdbalias atom ANE5AC O10 O
+pdbalias atom VCG C01 C1
+pdbalias atom VCG C02 C2
+pdbalias atom VCG C03 C3
+pdbalias atom VCG C04 C4
+pdbalias atom VCG C05 C5
+pdbalias atom VCG C06 C6
+pdbalias atom VCG C07 C7
+pdbalias atom VCG C08 C8
+pdbalias atom VCG C09 C9
+pdbalias atom TIP3 O OH2
+pdbalias atom CLA CL CLA
+pdbalias atom ACT CH3 C1
+pdbalias atom ACT O O1
+pdbalias atom ACT OXT O2
+pdbalias atom ACT C C2
+pdbalias atom CAL CA CAL
+pdbalias atom ILE CD1 CD
+pdbalias atom MET SE SD
+pdbalias atom BGLCNA C7 C
+pdbalias atom BGLCNA O7 O
+pdbalias atom BGLCNA C8 CT
+pdbalias atom BGLCNA N2 N
+pdbalias atom ANE5AC C10 C
+pdbalias atom ANE5AC C11 CT
+pdbalias atom ANE5AC N5 N
+pdbalias atom ANE5AC O1A O11
+pdbalias atom ANE5AC O1B O12
+pdbalias atom ANE5AC O10 O
+pdbalias atom VCG C01 C1
+pdbalias atom VCG C02 C2
+pdbalias atom VCG C03 C3
+pdbalias atom VCG C04 C4
+pdbalias atom VCG C05 C5
+pdbalias atom VCG C06 C6
+pdbalias atom VCG C07 C7
+pdbalias atom VCG C08 C8
+pdbalias atom VCG C09 C9
+pdbalias atom TIP3 O OH2
+pdbalias atom CLA CL CLA
+pdbalias atom ACT CH3 C1
+pdbalias atom ACT O O1
+pdbalias atom ACT OXT O2
+pdbalias atom ACT C C2
+pdbalias atom CAL CA CAL
+pdbalias residue HIS HSD
+pdbalias residue MSE MET
+pdbalias residue PO4 H2PO4
+pdbalias residue H2PO H2PO4
+pdbalias residue MAN AMAN
+pdbalias residue BMA BMAN
+pdbalias residue BGLC BGLCNA
+pdbalias residue NAG BGLCNA
+pdbalias residue NDG AGLCNA
+pdbalias residue FUC AFUC
+pdbalias residue FUL BFUC
+pdbalias residue GAL BGAL
+pdbalias residue SIA ANE5AC
+pdbalias residue ANE5 ANE5AC
+pdbalias residue EIC LIN
+pdbalias residue HOH TIP3
+pdbalias residue ZN ZN2
+pdbalias residue CL CLA
+pdbalias residue C6DH C6DHPC
+pdbalias residue C7DH C7DHPC
+pdbalias residue DT THY
+pdbalias residue DA ADE
+pdbalias residue DC CYT
+pdbalias residue DG GUA
+pdbalias residue DU URA
+pdbalias residue HEM HEME
+pdbalias residue TOCL TOCL1
+pdbalias residue ACT ACET
+pdbalias residue CA CAL
+pdbalias residue HIS HSD
+pdbalias residue MSE MET
+pdbalias residue PO4 H2PO4
+pdbalias residue H2PO H2PO4
+pdbalias residue MAN AMAN
+pdbalias residue BMA BMAN
+pdbalias residue BGLC BGLCNA
+pdbalias residue NAG BGLCNA
+pdbalias residue NDG AGLCNA
+pdbalias residue FUC AFUC
+pdbalias residue FUL BFUC
+pdbalias residue GAL BGAL
+pdbalias residue SIA ANE5AC
+pdbalias residue ANE5 ANE5AC
+pdbalias residue EIC LIN
+pdbalias residue HOH TIP3
+pdbalias residue ZN ZN2
+pdbalias residue CL CLA
+pdbalias residue C6DH C6DHPC
+pdbalias residue C7DH C7DHPC
+pdbalias residue DT THY
+pdbalias residue DA ADE
+pdbalias residue DC CYT
+pdbalias residue DG GUA
+pdbalias residue DU URA
+pdbalias residue HEM HEME
+pdbalias residue TOCL TOCL1
+pdbalias residue ACT ACET
+pdbalias residue CA CAL
+pdbalias residue HIS HSD
+pdbalias residue MSE MET
+pdbalias residue PO4 H2PO4
+pdbalias residue H2PO H2PO4
+pdbalias residue MAN AMAN
+pdbalias residue BMA BMAN
+pdbalias residue BGLC BGLCNA
+pdbalias residue NAG BGLCNA
+pdbalias residue NDG AGLCNA
+pdbalias residue FUC AFUC
+pdbalias residue FUL BFUC
+pdbalias residue GAL BGAL
+pdbalias residue SIA ANE5AC
+pdbalias residue ANE5 ANE5AC
+pdbalias residue EIC LIN
+pdbalias residue HOH TIP3
+pdbalias residue ZN ZN2
+pdbalias residue CL CLA
+pdbalias residue C6DH C6DHPC
+pdbalias residue C7DH C7DHPC
+pdbalias residue DT THY
+pdbalias residue DA ADE
+pdbalias residue DC CYT
+pdbalias residue DG GUA
+pdbalias residue DU URA
+pdbalias residue HEM HEME
+pdbalias residue TOCL TOCL1
+pdbalias residue ACT ACET
+pdbalias residue CA CAL
+readpsf my_6pti.psf pdb my_6pti.pdb
+writepsf cmap 00-01-000_psfgen-build.psf
+writepdb 00-01-000_psfgen-build.pdb
+exit
+########################### END PESTIFER VMD SCRIPT ############################
+######################## Thank you for using Pestifer! #########################